+++ common/world-countries/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @cloud-dragon/world-countries
 
+## 0.8.4
+
+### Patch Changes
+
+- add media query and storage
+- Updated dependencies
+  - @cloud-dragon/common-utils@0.8.4
+
 ## 0.8.3
 
 ### Patch Changes