+++ design/components-rn/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @cloud-design/components-rn
 
+## 0.8.4
+
+### Patch Changes
+
+- add media query and storage
+- Updated dependencies
+  - @cloud-dragon/react-utils@0.8.4
+  - @cloud-dragon/common-utils@0.8.4
+  - @cloud-dragon/world-countries@0.8.4
+
 ## 0.8.3
 
 ### Patch Changes