+++ design/headless/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @cloud-design/headless
 
+## 0.8.4
+
+### Patch Changes
+
+- add media query and storage
+- Updated dependencies
+  - @cloud-dragon/react-utils@0.8.4
+
 ## 0.8.3
 
 ### Patch Changes