+++ obx/react/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @cloud-dragon/obx-react
 
+## 0.8.4
+
+### Patch Changes
+
+- add media query and storage
+- Updated dependencies
+  - @cloud-dragon/react-utils@0.8.4
+  - @cloud-dragon/obx-core@0.8.4
+
 ## 0.8.3
 
 ### Patch Changes